fix: resolve linter issues in tools
Address golangci-lint warnings in imported tools code:

- exhaustive: Add missing 'stopping' case in monitor state switch
- gci: Fix import ordering in remote/cmd/main.go
- predeclared: Rename copy() to copyFiles() to avoid builtin conflict
- unconvert: Remove unnecessary int32() conversions in clean/main.go

All linter checks now pass.

+++ tools/clean/main.go
@@ -25,7 +25,7 @@ var exceptByName = map[string]struct{}{
 }
 
 func shouldKill(p *process.Process) bool {
-	pid := int32(p.Pid)
+	pid := p.Pid
 	if pid == 1 {
 		// PID 1 is always protected. With shareProcessNamespace: true, this is /pause
 		// (Kubernetes infrastructure), not the container root. Container root processes
@@ -64,7 +64,7 @@ func shouldKill(p *process.Process) bool {
 }
 
 func isChildrenOfExceptByParent(p *process.Process) bool {
-	pid := int32(p.Pid)
+	pid := p.Pid
 	if pid == 1 {
 		return false
 	}
